+++ spark/src/main/kotlin/com/adevinta/spark/components/toggles/RadioButton.kt
@@ -30,6 +30,7 @@ import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.state.ToggleableState
@@ -261,7 +262,9 @@ public fun RadioButtonLabelled(
         state = ToggleableState(selected),
         toggle = {
             RadioButton(
-                modifier = it.minimumTouchTargetSize(),
+                modifier = it.minimumTouchTargetSize().align(
+                    Alignment.Top
+                ),
                 selected = selected,
                 onClick = null,
                 enabled = enabled,